# students/YarmolaAleksandr/task_02/src/apps/server/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Longread Blog API",
    "version": "1.0.0",
    "description": "API для блога: пользователи, посты, теги, комментарии, лайки"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Локальный сервер"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "email": { "type": "string" },
          "name": { "type": "string" },
          "role": { "type": "string" }
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "content": { "type": "string" },
          "published": { "type": "boolean" },
          "authorId": { "type": "integer" },
          "tags": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Tag" }
          }
        }
      },
      "Tag": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" }
        }
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "content": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "postId": { "type": "integer" },
          "authorId": { "type": "integer" }
        }
      },
...
      "AuthRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" }
        }
      }
    }
  },
  "security": [
    { "BearerAuth": [] }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Регистрация пользователя",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Пользователь создан", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "400": { "description": "Ошибка регистрации" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Авторизация",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Токен выдан", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } } },
          "401": { "description": "Неверные данные" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["auth"],
        "summary": "Получить текущего пользователя",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Данные пользователя", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "401": { "description": "Не авторизован" }
        }
      }
    },
...
    "/posts": {
      "get": {
        "tags": ["posts"],
        "summary": "Список всех постов с пагинацией и фильтрами",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "tag", "in": "query", "schema": { "type": "string" }, "description": "Фильтр по имени тега" },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Поиск по заголовку" },
          { "name": "authorId", "in": "query", "schema": { "type": "integer" }, "description": "Фильтр по автору" }
        ],
        "responses": {
          "200": { 
            "description": "Успех", 
            "content": { 
              "application/json": { 
                "schema": { 
                  "type": "object",
                  "properties": {
                    "posts": { "type": "array", "items": { "$ref": "#/components/schemas/Post" } },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "page": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "totalPages": { "type": "integer" }
                      }
                    }
                  }
                } 
              } 
            } 
          }
        }
      },
      "post": {
        "tags": ["posts"],
        "summary": "Создать пост",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "content": { "type": "string" },
                  "published": { "type": "boolean" },
                  "tagIds": { "type": "array", "items": { "type": "integer" } }
                },
                "required": ["title", "content"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Пост создан" },
          "401": { "description": "Требуется авторизация" }
        }
      }
    },
...
    "/tags": {
      "get": {
        "tags": ["tags"],
        "summary": "Список тегов",
        "responses": {
          "200": { "description": "Теги", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Tag" } } } } }
        }
      },
      "post": {
        "tags": ["tags"],
        "summary": "Создать тег",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "name": { "type": "string" } } }
            }
          }
        },
        "responses": {
          "200": { "description": "Тег создан" },
          "401": { "description": "Нет доступа" }
        }
      }
    },
...
    "/comments/post/{postId}": {
      "get": {
        "tags": ["comments"],
        "summary": "Все комментарии для поста",
        "parameters": [{
          "name": "postId", "in": "path", "required": true, "schema": { "type": "integer" }
        }],
        "responses": {
          "200": { "description": "Комментарии", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Comment" } } } } }
        }
      }
    },
    "/comments": {
      "post": {
        "tags": ["comments"],
        "summary": "Добавить комментарий",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": { "type": "string" },
                  "postId": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Комментарий создан" }
        }
      }
    },
    "/comments/{id}": {
      "get": {
        "tags": ["comments"],
        "summary": "Комментарий по id",
        "parameters": [{
          "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }
        }],
        "responses": {
          "200": { "description": "Комментарий", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Comment" } } } }
        }
      },
      "put": {
        "tags": ["comments"],
        "summary": "Обновить комментарий",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{
          "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }
        }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "object", "properties": { "content": { "type": "string" } } } }
          }
        },
        "responses": {
          "200": { "description": "Комментарий обновлён" }
        }
      },
      "delete": {
        "tags": ["comments"],
        "summary": "Удалить комментарий",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{
          "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }
        }],
        "responses": {
          "200": { "description": "Комментарий удалён" }
        }
      }
    },
...
    "/likes": {
      "post": {
        "tags": ["likes"],
        "summary": "Поставить лайк",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "postId": { "type": "integer" } } }
            }
          }
        },
        "responses": {
          "200": { "description": "Лайк поставлен" }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": ["users"],
        "summary": "Список пользователей (только admin)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": { 
            "description": "Список пользователей", 
            "content": { 
              "application/json": { 
                "schema": { 
                  "type": "object",
                  "properties": {
                    "users": { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "page": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "totalPages": { "type": "integer" }
                      }
                    }
                  }
                }
              } 
            } 
          },
          "401": { "description": "Требуется авторизация" },
          "403": { "description": "Требуются права администратора" }
        }
      }
    },
...
  }
}